add theme to pub listings
